Feat : root 프로젝트 exception, util 패키지 module-core로 이동
