+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/exceptions/exception/PasswordMismatchException.java
@@ -1,0 +1,11 @@
+package com.recipe.jamanchu.core.exceptions.exception;
+
+import com.recipe.jamanchu.core.exceptions.ExceptionStatus;
+
+public class PasswordMismatchException extends GlobalException {
+
+  public PasswordMismatchException() {
+
+    super(ExceptionStatus.PASSWORD_MISMATCH);
+  }
+}
