+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/exceptions/exception/RecipeNotFoundException.java
@@ -1,0 +1,10 @@
+package com.recipe.jamanchu.core.exceptions.exception;
+
+import com.recipe.jamanchu.core.exceptions.ExceptionStatus;
+
+public class RecipeNotFoundException extends GlobalException {
+
+  public RecipeNotFoundException() {
+    super(ExceptionStatus.RECIPE_NOT_FOUND);
+  }
+}
