+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/exceptions/exception/SocialAccountException.java
@@ -1,0 +1,11 @@
+package com.recipe.jamanchu.core.exceptions.exception;
+
+import com.recipe.jamanchu.core.exceptions.ExceptionStatus;
+
+public class SocialAccountException extends GlobalException {
+
+  public SocialAccountException() {
+
+    super(ExceptionStatus.SOCIAL_ACCOUNT);
+  }
+}
