+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/exceptions/exception/UserInfoRetrievalException.java
@@ -1,0 +1,12 @@
+package com.recipe.jamanchu.core.exceptions.exception;
+
+import com.recipe.jamanchu.core.exceptions.ExceptionStatus;
+
+public class UserInfoRetrievalException extends GlobalException {
+
+  public UserInfoRetrievalException() {
+
+    super(ExceptionStatus.USER_INFO_RETRIEVAL);
+  }
+
+}
