+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/exceptions/exception/UserNotFoundException.java
@@ -1,0 +1,11 @@
+package com.recipe.jamanchu.core.exceptions.exception;
+
+import com.recipe.jamanchu.core.exceptions.ExceptionStatus;
+
+public class UserNotFoundException extends GlobalException {
+
+  public UserNotFoundException() {
+
+    super(ExceptionStatus.USER_NOT_FOUND);
+  }
+}
