+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/util/PictureManager.java
@@ -1,11 +1,10 @@
-package com.recipe.jamanchu.util;
+package com.recipe.jamanchu.core.util;
 
 import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.AmazonS3Exception;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.recipe.jamanchu.model.type.PictureType;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
