+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/util/PictureType.java
@@ -1,0 +1,13 @@
+package com.recipe.jamanchu.core.util;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum PictureType {
+  THUMBNAIL("thumbnail"),
+  RECIPE_ORDER_IMAGE("recipe_order_image");
+
+  private final String folderPrefix;
+}
