+++ jamanchu/module-core/src/main/java/com/recipe/jamanchu/core/util/RecipeIngredientMappingId.java
@@ -1,4 +1,4 @@
-package com.recipe.jamanchu.util;
+package com.recipe.jamanchu.core.util;
 
 import java.io.Serializable;
 import java.util.Objects;
