--- jamanchu/src/main/java/com/recipe/jamanchu/exceptions/exception/PasswordMismatchException.java
@@ -1,11 +1,0 @@
-package com.recipe.jamanchu.exceptions.exception;
-
-import com.recipe.jamanchu.exceptions.ExceptionStatus;
-
-public class PasswordMismatchException extends GlobalException{
-
-  public PasswordMismatchException() {
-
-    super(ExceptionStatus.PASSWORD_MISMATCH);
-  }
-}
