--- jamanchu/src/main/java/com/recipe/jamanchu/exceptions/exception/SocialAccountException.java
@@ -1,11 +1,0 @@
-package com.recipe.jamanchu.exceptions.exception;
-
-import com.recipe.jamanchu.exceptions.ExceptionStatus;
-
-public class SocialAccountException extends GlobalException{
-
-  public SocialAccountException() {
-
-    super(ExceptionStatus.SOCIAL_ACCOUNT);
-  }
-}
