--- jamanchu/src/main/java/com/recipe/jamanchu/exceptions/exception/UserInfoRetrievalException.java
@@ -1,12 +1,0 @@
-package com.recipe.jamanchu.exceptions.exception;
-
-import com.recipe.jamanchu.exceptions.ExceptionStatus;
-
-public class UserInfoRetrievalException extends GlobalException{
-
-  public UserInfoRetrievalException() {
-
-    super(ExceptionStatus.USER_INFO_RETRIEVAL);
-  }
-
-}
