--- jamanchu/src/main/java/com/recipe/jamanchu/exceptions/exception/UserNotFoundException.java
@@ -1,11 +1,0 @@
-package com.recipe.jamanchu.exceptions.exception;
-
-import com.recipe.jamanchu.exceptions.ExceptionStatus;
-
-public class UserNotFoundException extends GlobalException{
-
-  public UserNotFoundException() {
-
-    super(ExceptionStatus.USER_NOT_FOUND);
-  }
-}
